+++ backend/src/models/Employers/deleteEmployerByID.js
@@ -2,7 +2,7 @@ import supabase from "../../config/supabase.js";
 
 const deleteEmployerByID = async (id) => {
     try {
-        const { error } = await supabase.from("users").delete().eq("id", id);
+        const { error } = await supabase.from("users").delete().eq("id", id).eq("role", 1);
         if (error) return error;
     }
     catch (err) {
